Fix Merge Conflicts & Typing
