+++ .env
@@ -6,4 +6,4 @@ VITE_FIREBASE_MESSAGINGSENDERID='91762702683'
 VITE_FIREBASE_APPID='1:91762702683:web:d4d49322d9dfe1034e427e'
 
 VITE_FINAGE_API_KEY='API_KEY761N7UUP1D7HE2S27YUMGE8BHWBGU7J5'
-VITE_STOCKNEWS_API_KEY='pcuyi0cz4u7kw7aisqtc47vonwafyz0a8emd1xst'
+VITE_STOCK_NEWS_API_KEY='pcuyi0cz4u7kw7aisqtc47vonwafyz0a8emd1xst'
